backend/requirements.txt
flask
google-api-python-client
openai
youtube-transcript-api
google-cloud-vision
